+++ DotNetHowToSecurityTrimming/DotNetHowToSecurityTrimming/App.config
@@ -4,14 +4,10 @@
     <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.1" />
   </startup>
   <appSettings>
-    <!--<add key="ClientId" value="[Your_Client_Id]" />
+    <add key="ClientId" value="[Your_Client_Id]" />
     <add key="Tenant" value="[Your_Tenant]" />
     <add key="SearchServiceApiKey" value="[Your_API_Key]" />
-    <add key="SearchServiceName" value="[Your_Service_Name]" />-->
-   <add key="ClientId" value="a1b6ff66-d087-42dd-af99-7dab5b43e56b"/>
-    <add key="Tenant" value="revitals.onmicrosoft.com" />
-    <add key="SearchServiceApiKey" value="8D549882F7E65DB0829E4F9A8C25ED67" />
-    <add key="SearchServiceName" value="revitalbtest" />
+    <add key="SearchServiceName" value="[Your_Service_Name]" />
   </appSettings>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
