--- DotNetHowToSecurityTrimming/DotNetHowToSecurityTrimming/Contract/Response.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace DotNetHowToSecurityTrimming
-{
-    public class Response
-    {
-        public string Id { get; set; }
-
-        public ResponseBody Body { get; set; }
-    }
-}
