--- DotNetHowToSecurityTrimming/DotNetHowToSecurityTrimming/Contract/ResponseBody.cs
@@ -1,14 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DotNetHowToSecurityTrimming
-{
-    public class ResponseBody
-    {
-        public ResponseBodyValue[] Value { get; set; }
-    }
-}
