--- DotNetHowToSecurityTrimming/DotNetHowToSecurityTrimming/Contract/Result.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace DotNetHowToSecurityTrimming
-{
-    public class Result
-    {
-        public Response[] Responses { get; set; }
-    }
-}
